third_party/blink/web_tests/media/media-playback-while-not-visible-permission-policy-pause-when-hidden-nested.html

<!DOCTYPE html>
<title>Test behavior of HTMLMediaElements with the media-playback-while-not-rendered permission policy in nested iframes</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<script>
  internals.settings.setAutoplayPolicy('no-user-gesture-required');

  function hideFrame(iframe, type) {
    if (type === 'display') {
      iframe.style.setProperty('display', 'none');
    } else if (type === 'visibility') {
      iframe.style.setProperty('visibility', 'hidden');
    } else if (type === 'zeroSize') {
      iframe.style.setProperty('width', '0');
      iframe.style.setProperty('height', '0');
    }
  }

  function runPauseTest(mediaTag, mediaSrc, type) {
    async_test(function(t) {
      const intermediateIframe = document.createElement('iframe');
      intermediateIframe.allow = "media-playback-while-not-visible 'none'; autoplay *";
      intermediateIframe.srcdoc =
          `<!DOCTYPE html><html><body><iframe id='nested-iframe' srcdoc='<!DOCTYPE html><${mediaTag} autoplay loop></${mediaTag}>'></iframe></body></html>`;
      document.body.appendChild(intermediateIframe);
      t.add_cleanup(() => { document.body.removeChild(intermediateIframe); });

      intermediateIframe.onload = t.step_func(function() {
        const nestedIframe =
            intermediateIframe.contentDocument.getElementById('nested-iframe');
        const media = nestedIframe.contentDocument.querySelector(mediaTag);
        assert_true(media.paused);

        media.addEventListener('playing', t.step_func(function() {
          assert_false(media.paused);
          // Hiding the intermediate iframe should pause the embedded media
          // element playback in the nested iframe.
          hideFrame(intermediateIframe, type);

          media.addEventListener('pause', t.step_func(function() {
            assert_true(media.paused);
            media.play().then(
                t.unreached_func("Unrendered media element should not be allowed to play."),
                (e) => {
                  assert_true(e.name == 'NotAllowedError');
                  t.done();
                });
          }));
        }));

        media.src = mediaSrc;
      });
    }, `${mediaTag} playback should pause for nested hidden iframes when media-playback-while-not-visible permission policy is set (type = ${type})`);
  }

  const visibility_types = ['display', 'visibility', 'zeroSize'];
  for (const type of visibility_types) {
    runPauseTest('video', 'content/test.ogv', type);
    runPauseTest('audio', 'content/test.oga', type);
  }
</script>
</body>
